+++ packages/components-js/tests/vrt/puppeteer/specs/component-high-contrast-mode.vrt.ts
@@ -50,6 +50,7 @@ const components: Component[] = [
   'text-list',
   'textarea-wrapper',
   'toast-basic',
+  'wordmark',
 ];
 
 const scenarioPopover = async (page: Page) => {