+++ packages/storefront/src/components/CodeBlock.vue
@@ -156,13 +156,13 @@
             .token.char,
             .token.builtin,
             .token.inserted {
-              color: #a55a00;
+              color: #7e4603;
             }
 
             .token.atrule,
             .token.attr-value,
             .token.function {
-              color: #3d7b3c;
+              color: #30612f;
             }
           }
         }